pyproject.toml
[project]
name = "com-chatmoshi-base"
version = "23.10.0"
description = "Lightweight tools used across all ChatMoshi components."
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = {text = "Proprietary"}
keywords = ["language", "ai", "moshi", "base", "types", "storage"]
authors = [
  {name = "Eric Kalosa-Kenyon", email = "moshi.feedback@gmail.com"}
]
dependencies = [
  "iso639",
  "isocodes",
  "google-cloud-firestore",
  "langcodes[data]",
  "loguru",
  "openai",
]

[project.optional-dependencies]
test = [
  "pytest",
  "pytest-cov",
]
dev = [
  "black",
  "flake8",
  "isort",
  "pyfiglet"
]

[project.urls]
Homepage = "https://www.chatmoshi.com"
Repository = "https://github.com/ekalosak/moshi-base.git"


[tool.pytest.ini_options]
markers = [
  "slow: runs more than a second or two to run",
  "fb: requires an emulated firebase instance to run",
]
addopts = [
  "--cov=moshi"
]
